requirements.txt
duckdb>=0.10.0
snowflake-connector-python>=3.6.0
streamlit>=1.30.0
pyyaml>=6.0
pandas>=2.0.0
pyarrow>=14.0.0
apscheduler>=3.10.0
watchdog>=3.0.0
dbt-core>=1.7.0
dbt-duckdb>=1.7.0
